.windsurf/skills/dtx/dtx-state/scripts/include/pyproject.toml
[project]
name = "dtx-state-cli"
version = "0.1.0"
description = "CLI for dtx-state skill"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
